fix: Split getParams functions
