+++ src/v2/Utils/getServerParam.tsx
@@ -1,0 +1,6 @@
+export function getServerParam(req, name): string | null {
+  if (req.query && req.query.hasOwnProperty(name)) {
+    return req.query[name]
+  }
+  return null
+}
